requirements.txt
Django==1.10.1
django-mptt==0.8.6
gunicorn==19.6.0
oauthlib==2.0.0
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
whitenoise==3.2.1
